# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API сервиса многостороннего обмена вещами: профили пользователей, вход по JWT\nи объявления о вещах, которые владелец готов обменять.\n\nЗащищённые маршруты читают HttpOnly cookie `access_token` — они помечены замком.\nКнопки «Authorize» здесь нет и не нужно: выполните `POST /auth/login` прямо из этой\nстраницы — браузер сохранит cookie и будет отправлять её со всеми следующими\nзапросами сам. JavaScript до неё не дотянется, поэтому вписать её руками нельзя.",
        "title": "Цепочка обмена — API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Справочник для полей `category` и `wants`. Слаг — стабильный ключ, название меняется, слаг нет.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Список категорий",
                "responses": {
                    "200": {
                        "description": "Категории",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_item_dto.CategoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            }
        },
...
        "/items": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Требует cookie `access_token`. Возвращает все объявления текущего пользователя\nлюбого статуса, от новых к старым. Чужие объявления этим маршрутом не отдаются.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Получить свои объявления",
                "responses": {
                    "200": {
                        "description": "Список объявлений; если их нет, возвращается []",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ItemResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Нет или истекла cookie access_token",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Требует cookie `access_token`. Владелец берётся из токена, а не из тела запроса.\nНужна хотя бы одна фотография (ссылкой) и хотя бы одна желаемая категория —\nбез них объявление не участвует в подборе обменов. Список категорий: GET /categories.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Создать объявление",
                "parameters": [
                    {
                        "description": "Данные объявления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Создано, ссылка на объявление в заголовке Location",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса, нет фото или желаний, неизвестная категория",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "401": {
                        "description": "Нет или истекла cookie access_token",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Публичная карточка вещи: фотографии, желаемые категории и статус. Аутентификация не нужна.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Получить объявление по ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "8db9f3e2-8a45-4a70-b3d1-167b4f97e121",
                        "description": "UUID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объявление",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemResponse"
                        }
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "404": {
                        "description": "Объявление не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Требует cookie `access_token`, удалять можно только свои объявления. Вещь, уже занятую\nв незавершённом обмене, удалить нельзя: иначе участник остался бы без обещанного предмета.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Удалить своё объявление",
                "parameters": [
                    {
                        "type": "string",
                        "example": "8db9f3e2-8a45-4a70-b3d1-167b4f97e121",
                        "description": "UUID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "401": {
                        "description": "Нет или истекла cookie access_token",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "403": {
                        "description": "Объявление принадлежит другому пользователю",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "404": {
                        "description": "Объявление не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "409": {
                        "description": "Вещь участвует в незавершённом обмене",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Требует cookie `access_token`, менять можно только свои объявления. Достаточно одного поля.\n`photo_urls` и `wants` заменяются целиком: чтобы добавить фотографию, пришлите старые\nссылки вместе с новой. Пустой список запрещён — у объявления всегда есть хотя бы одно фото.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Изменить своё объявление",
                "parameters": [
                    {
                        "type": "string",
                        "example": "8db9f3e2-8a45-4a70-b3d1-167b4f97e121",
                        "description": "UUID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля, которые нужно изменить",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённое объявление",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело, пустой список фото или желаний, неизвестная категория",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "401": {
                        "description": "Нет или истекла cookie access_token",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "403": {
                        "description": "Объявление принадлежит другому пользователю",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "404": {
                        "description": "Объявление не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "409": {
                        "description": "Нельзя изменить условия объявления в незавершённом обмене",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ItemError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateItemRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "bikes"
                },
                "description": {
                    "type": "string",
                    "example": "Почти новый, катался год"
                },
                "photo_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/bike-1.jpg",
                        "https://example.com/bike-2.jpg"
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "Велосипед"
                },
                "wants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "consoles",
                        "phones"
                    ]
                }
            }
        },
...
        "dto.ItemError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.ItemPickupPoint": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ItemResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "photo_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pickup_point": {
                    "description": "null — вещь дома у владельца.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ItemPickupPoint"
                        }
                    ],
                    "x-nullable": true
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_sweetlife999_chain-of-trades-avito_internal_item_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Велосипеды и транспорт"
                },
                "slug": {
                    "type": "string",
                    "example": "bikes"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "HttpOnly cookie, которую ставит POST /auth/login.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}